Rename `message` module to be `messages`
